Main Menu Working
I have gotten the main manu to kinda work.
However now when you press play the game will hang then crash. Still it's an improvement from before.

I had to do a lot of hacky work around to get this to work, so improvements will be necessary to make this more robust.